+++ locale/ar_MA.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * eduVPN - End-user friendly VPN.
+ *
+ * Copyright: 2016-2020, The Commons Conservancy eduVPN Programme
+ * SPDX-License-Identifier: AGPL-3.0+
+ */
+
+return [
+    'Android' => 'Android',
+    'Continue...' => 'واصل...',
+    'Linux' => 'Linux',
+    'On the "Configurations" page, advanced users can manually download VPN configurations.' => 'على صفحة "الاعدادات"، يمكن للمستعملين المتمرسين تحميل اعدادات الشبكة الخاصة الافتراضية يدويا.',
+    'On your "Account" page you can block access to the VPN in case you lose a device.' => 'على صفحة "حساب" تستطيع منع الولوج الى الشبكة الخاصة الافتراضية اذا فقدت جهازا ما.',
+    'To use Let\'s Connect!, download the app for your device below!' => 'لاستخدام Let\'s Connect!، حمل التطبيق لجهازك اسفله!',
+    'Welcome to Let\'s Connect!' => 'مرحبا بك على Let\'s Connect!',
+    'Windows' => 'Windows',
+    'iOS' => 'iOS',
+    'macOS' => 'macOS',
+];